requirements.txt
omegaconf
jieba_fast
pypinyin
opencc
transformers
sentencepiece
onnxruntime
pytorch_lightning
einops
tqdm
librosa
scipy
ffmpeg-python
soundfile
g2p_en
LangSegment
wordsegment
nltk
pyopenjtalk
vector_quantize_pytorch
loralib
einops